moving network host tests to their own class
