+++ Assets/Mirror/Tests/Editor/NetworkHostTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 386969046dd50464aa936377258de6be
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
